storyline5_pipeline/requirements.txt
pandas>=1.5.0
openpyxl>=3.1.0
numpy>=1.24.0
matplotlib>=3.7.0
PyYAML>=6.0
